+++ src/background/modules/mod.rs
@@ -1,3 +1,4 @@
+pub mod apps;
 pub mod bluetooth;
 pub mod input;
 pub mod language;
@@ -9,7 +10,6 @@ pub mod power;
 pub mod shared;
 pub mod start;
 pub mod system_settings;
-pub mod tray;
 pub mod user;
 pub mod uwp;