+++ src/background/system/mod.rs
@@ -5,6 +5,7 @@ use crate::{
     error::Result,
     log_error,
     modules::{
+        apps::infrastructure::register_app_win_events,
         bluetooth::{infrastructure::register_bluetooth_events, release_bluetooth_events},
         language::register_language_events,
         media::infrastructure::{register_media_events, release_media_events},
@@ -16,7 +17,6 @@ use crate::{
         power::infrastructure::{register_power_events, release_power_events},
         shared::radio::RADIO_MANAGER,
         system_settings::infrastructure::{register_system_settings_events, release_colors_events},
-        tray::infrastructure::register_tray_icons_events,
         user::infrastructure::register_user_events,
     },
     trace_lock,
@@ -38,7 +38,7 @@ pub fn declare_system_events_handlers() -> Result<()> {
         log_error!(register_network_events());
     });
 
-    register_tray_icons_events();
+    register_app_win_events();
     register_notification_events();
     register_media_events();
     register_user_events();