+++ src/background/widgets/window_manager/mod.rs
@@ -35,14 +35,15 @@ impl WindowManagerV2 {
         let exe = window.process().program_path();
 
         if let Ok(exe) = &exe {
-            if exe.ends_with("ApplicationFrameHost.exe") && window.is_real_window() {
+            if exe.ends_with("ApplicationFrameHost.exe") && window.is_interactable_and_not_hidden()
+            {
                 return true;
             }
         }
 
         // Without admin some apps does not return the exe path so these should be unmanaged
         exe.is_ok()
-        && window.is_real_window()
+        && window.is_interactable_and_not_hidden()
         // Ignore windows without a title bar, and top most windows normally are widgets or tools so they should not be managed
         && (WindowsApi::get_styles(hwnd).contains(WS_CAPTION) && !WindowsApi::get_ex_styles(hwnd).contains(WS_EX_TOPMOST))
         && !window.is_cloaked()
@@ -88,7 +89,7 @@ impl WindowManagerV2 {
     }
 
     fn add(window: &Window) -> Result<()> {
-        let monitor_id = window.get_cached_data().monitor;
+        let monitor_id = window.monitor_id();
 
         let mut vd_manager = get_vd_manager();
         let vd_active_id = vd_manager.get_active_workspace_id(&monitor_id).clone();
@@ -108,7 +109,7 @@ impl WindowManagerV2 {
     }
 
     fn remove(window: &Window) -> Result<()> {
-        let monitor_id = window.get_cached_data().monitor;
+        let monitor_id = window.monitor_id();
         let mut state = trace_lock!(WM_STATE);
 
         let mut vd = get_vd_manager();